// src/app/pages/PersonalProposal/credit-info/Models/deposite-in-bank.ts
export class DepositeInBank {
    ID: number;
    CLIENT_ID: number;
    CREDIT_INFORMATION_ID: number;
    BANK_NAME: string;
    BRANCH_NAME: string;
    DEPOSIT_TYPE: string;
    DEPOSITE_AMOUNT: number;
    IS_LOAN_TAKEN: boolean;
    IS_SUB: boolean;
    DEPOSITE_TYPE: string;
    ACCOUNT_NO:number
}